requirements.txt
gradio
numpy
openai
opencv_python
Pillow
shortuuid
spacy
supervision
torch
torchvision
tqdm
transformers
bitsandbytes
accelerate